fix(Search): rewrite to TS, fix open zone. fix token/pair sorting, add name to pair interface
