+++ src/data/mappers/ethereum/pairMappers.ts
@@ -8,6 +8,7 @@ export function pairMapper(payload: EthereumPair, ethPrice: number): Pair {
     tokenOne: {
       id: payload.token0?.id || '',
       symbol: parseTokenInfo('symbol', payload.token0?.id, payload.token0?.symbol),
+      name: parseTokenInfo('name', payload.token0?.id, payload.token0?.name),
       reserve: payload.reserve0 ? +payload.reserve0 : 0,
       price: calculateTokenPrice(payload?.reserve1, payload?.reserve0),
       priceUSD: (payload.token0.derivedETH || 0) * ethPrice
@@ -15,6 +16,7 @@ export function pairMapper(payload: EthereumPair, ethPrice: number): Pair {
     tokenTwo: {
       id: payload.token1?.id || '',
       symbol: parseTokenInfo('symbol', payload.token1?.id, payload.token1?.symbol),
+      name: parseTokenInfo('name', payload.token1?.id, payload.token1?.name),
       reserve: payload.reserve1 ? +payload.reserve1 : 0,
       price: calculateTokenPrice(payload?.reserve0, payload?.reserve1),
       priceUSD: (payload.token1.derivedETH || 0) * ethPrice
